refactor: Externalized JWKS logic to a separated file to improve readability
