+++ pkg/oauth/jwk/jwk.go
@@ -89,14 +89,6 @@ func (key *JSONWebKey) RSA() (*rsa.PublicKey, error) {
 }
 
 /*
-JSONWebKeySet - Represents a list of public keys that can be used for validating token signatures
-*/
-type JSONWebKeySet struct {
-	// Keys - All Keys available for signing under the set
-	Keys []JSONWebKey `json:"keys" bson:"keys"`
-}
-
-/*
 New - Generates a new key depending on the algorithm that you specify in the parameter. Calling this function will
 immediately set the key as the current one, however this will not retroactively update previously issued key. If you are
 attempting to rotate/revoke keys, then you should use RotateKeys or RotateRevokeKeys.
@@ -184,41 +176,4 @@ func ActiveKey(serv *server.Server, alg string, audience string) (*PrivateJSONWe
 	}
 
 	return &jwk, nil
-}
-
-/*
-JWKS - Fetches all JSON Web Keys stored in the database and returns them as a slice. Only RSA Keys are returned with
-this function call, as this is intended to be used with the .well-known/jwks.json endpoint, and HSA secrets should not
-be exposed publicly as they are symmetrical
-
-TODO: Maybe rethink this to return only keys by a specific audience
-*/
-func JWKS(serv *server.Server) (*JSONWebKeySet, error) {
-	jwks := new(JSONWebKeySet)
-
-	/*
-		This function call is actually fairly simple, as all we really need to do here is list out the entire collection.
-	*/
-	cursor, err := serv.Database().Collection("jwk").Find(context.Background(), bson.M{"kty": "RSA"})
-	if err != nil {
-		if !errors.Is(err, mongo.ErrNoDocuments) && err != nil {
-			return nil, fmt.Errorf("%w (%v)", server.ErrInternalDatabase, err)
-		}
-	}
-
-	/*
-		Then we simply just decode all the results into our slice and then return it.
-	*/
-	err = cursor.All(context.Background(), &jwks.Keys) // check here for proper errors
-	if err != nil {
-		if !errors.Is(err, mongo.ErrNoDocuments) && err != nil {
-			return nil, fmt.Errorf("%w (%v)", server.ErrInternalDatabase, err)
-		}
-
-		if errors.Is(err, mongo.ErrNoDocuments) {
-			return nil, ErrKeyNotExist
-		}
-	}
-
-	return jwks, nil
 }
